refactor: floating ui hook
